% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absences}
\alias{absences}
\title{absences.}
\format{
A data frame with 977 rows and 6 variables:
\describe{
  \item{Lat}{latitude of sample}
  \item{Long}{longitude of sample}
  \item{Season}{One of four seasons}
  \item{Taxon}{Taxon that we are plotting}
  \item{freqfac}{frequency of occurence}
  \item{freqsamp}{frequency of occurence in no of sample}
  ...
}
}
\usage{
absences
}
\description{
A dataset containing the absences in the data for plotting.
}
\keyword{datasets}
