% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioregion}
\alias{bioregion}
\title{bioregion}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 11 rows and 6 columns.
}
\usage{
bioregion
}
\description{
A dataset containing the bioregion information.
}
\keyword{datasets}
