% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datCPRzts}
\alias{datCPRzts}
\title{datCPRzts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3800 rows and 7 columns.
}
\usage{
datCPRzts
}
\description{
A dataset containing the CPR time series information for plotting.
}
\keyword{datasets}
