% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datNRSi}
\alias{datNRSi}
\title{datNRSi}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7830 rows and 8 columns.
}
\usage{
datNRSi
}
\description{
A dataset containing the NRS time series information for plotting.
}
\keyword{datasets}
