% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_sf}
\alias{meta_sf}
\title{meta_sf}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.
}
\usage{
meta_sf
}
\description{
A dataset to turn the sf info into.
}
\keyword{datasets}
