% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRSBGCEnvData}
\alias{NRSBGCEnvData}
\title{NRSBGCEnvData}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96719 rows and 18 columns.
}
\usage{
NRSBGCEnvData
}
\description{
A dataset containing the environmental data for plotting.
}
\keyword{datasets}
