% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs}
\alias{obs}
\title{obs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 61280 rows and 5 columns.
}
\usage{
obs
}
\description{
A dataset containing the observations for plotting.
}
\keyword{datasets}
