% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Samples}
\alias{Samples}
\title{Samples}
\format{
An object of class \code{data.frame} with 977 rows and 4 columns.
}
\usage{
Samples
}
\description{
A dataset containing the samples used for plotting.
}
\keyword{datasets}
