% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampLocs}
\alias{SampLocs}
\title{SampLocs}
\format{
An object of class \code{data.frame} with 7970 rows and 5 columns.
}
\usage{
SampLocs
}
\description{
A dataset containing the sample locations used for plotting.
}
\keyword{datasets}
